refactor: Remove diff functionality
Why? It doesn't work the way we expect. Consider the following case (thanks to Lea Verou): cell had 1, Mavo loads the 1, then someone edits the sheet and changes the 1 to a 2, I save the Mavo without editing that cell, should the cell revert to 1 or remain 2? It should revert to 1. But for now, it's not.

Mark these words well: first design, then implement!