FIx issues with saving collections
The problem was that previously, for headings, I used the first element in the data returned by Mavo, but what I really needed to do was to use the real property names of an element of a collection. Luckily Mavo knows them.

Another issue was that elements of a simple collection could have a different shape if their values partly came from a spreadsheet and partly from the app: some of them might be objects and some primitives.